app/models/book.py
from app import db

class Book(db.Model):
    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
    title = db.Column(db.String)
    description = db.Column(db.String)

#An instance function that returns a dictionay which represents the book model
    def to_dict(self):
        return {
            'id': self.id,
            'title': self.title,
            'description': self.description
        }
@classmethod
def from_dict(cls, data):
    if 'title' not in data or 'description' not in data:
        raise ValueError("Missing required fields: title or description")
    return cls(title=data['title'], description=data['description'])

